+++ lib/src/dart_formatter.dart
@@ -25,6 +25,22 @@ import 'string_compare.dart' as string_compare;
 
 /// Dart source code formatter.
 class DartFormatter {
+  /// The latest Dart language version that can be parsed and formatted by this
+  /// version of the formatter.
+  static final latestLanguageVersion = Version(3, 3, 0);
+
+  /// The highest Dart language version without support for patterns.
+  static final _lastNonPatternsVersion = Version(2, 19, 0);
+
+  /// The Dart language version that formatted code should be parsed as.
+  ///
+  /// Note that a `// @dart=` comment inside the code overrides this.
+  final Version languageVersion;
+
+  /// Whether the user passed in a non-`null` language version.
+  // TODO(rnystrom): Remove this when the language version is required.
+  final bool _omittedLanguageVersion;
+
   /// The string that newlines should use.
   ///
   /// If not explicitly provided, this is inferred from the source text. If the
@@ -45,7 +61,15 @@ class DartFormatter {
   /// See dart.dev/go/experiments for details.
   final List<String> experimentFlags;
 
-  /// Creates a new formatter for Dart code.
+  /// Creates a new formatter for Dart code at [languageVersion].
+  ///
+  /// If [languageVersion] is omitted, then it defaults to
+  /// [latestLanguageVersion]. In a future major release of dart_style, the
+  /// language version will affect the applied formatting style. At that point,
+  /// this parameter will become required so that the applied style doesn't
+  /// change unexpectedly. It is optional now so that users can migrate to
+  /// versions of dart_style that accept this parameter and be ready for the
+  /// major version when it's released.
   ///
   /// If [lineEnding] is given, that will be used for any newlines in the
   /// output. Otherwise, the line separator will be inferred from the line
@@ -56,12 +80,15 @@ class DartFormatter {
   ///
   /// While formatting, also applies any of the given [fixes].
   DartFormatter(
-      {this.lineEnding,
+      {Version? languageVersion,
+      this.lineEnding,
       int? pageWidth,
       int? indent,
       Iterable<StyleFix>? fixes,
       List<String>? experimentFlags})
-      : pageWidth = pageWidth ?? 80,
+      : languageVersion = languageVersion ?? latestLanguageVersion,
+        _omittedLanguageVersion = languageVersion == null,
+        pageWidth = pageWidth ?? 80,
         indent = indent ?? 0,
         fixes = {...?fixes},
         experimentFlags = [...?experimentFlags];
@@ -72,18 +99,15 @@ class DartFormatter {
   /// If [uri] is given, it is a [String] or [Uri] used to identify the file
   /// being formatted in error messages.
   String format(String source, {Object? uri}) {
-    if (uri == null) {
-      // Do nothing.
-    } else if (uri is Uri) {
-      uri = uri.toString();
-    } else if (uri is String) {
-      // Do nothing.
-    } else {
-      throw ArgumentError('uri must be `null`, a Uri, or a String.');
-    }
+    var uriString = switch (uri) {
+      null => null,
+      Uri() => uri.toString(),
+      String() => uri,
+      _ => throw ArgumentError('uri must be `null`, a Uri, or a String.'),
+    };
 
     return formatSource(
-            SourceCode(source, uri: uri as String?, isCompilationUnit: true))
+            SourceCode(source, uri: uriString, isCompilationUnit: true))
         .text;
   }
 
@@ -118,13 +142,26 @@ class DartFormatter {
     }
 
     // Parse it.
-    var parseResult = _parse(text, source.uri, patterns: true);
+    var parseResult = _parse(text, source.uri, languageVersion);
 
-    // If we couldn't parse it with patterns enabled, it may be because of
-    // one of the breaking syntax changes to switch cases. Try parsing it
-    // again without patterns.
-    if (parseResult.errors.isNotEmpty) {
-      var withoutPatternsResult = _parse(text, source.uri, patterns: false);
+    // If we couldn't parse it, and the language version supports patterns, it
+    // may be because of the breaking syntax changes to switch cases. Try
+    // parsing it again without pattern support.
+    // TODO(rnystrom): This is a pretty big hack. Before Dart 3.0, every
+    // language version was a strict syntactic superset of all previous
+    // versions. When patterns were added, a small number of switch cases
+    // became syntax errors.
+    //
+    // For most of its history, the formatter simply parsed every file at the
+    // latest language version without having to detect each file's actual
+    // version. We are moving towards requiring the language version when
+    // formatting, but for now, try to degrade gracefully if the user omits the
+    // version.
+    //
+    // Remove this when the languageVersion constructor parameter is required.
+    if (_omittedLanguageVersion && parseResult.errors.isNotEmpty) {
+      var withoutPatternsResult =
+          _parse(text, source.uri, _lastNonPatternsVersion);
 
       // If we succeeded this time, use this parse instead.
       if (withoutPatternsResult.errors.isEmpty) {
@@ -197,31 +234,8 @@ class DartFormatter {
     return output;
   }
 
-  /// Parse [source] from [uri].
-  ///
-  /// If [patterns] is `true`, the parse at the latest language version
-  /// which supports patterns and treats switch cases as patterns. If `false`,
-  /// then parses using an older language version where switch cases are
-  /// constant expressions.
-  ///
-  // TODO(rnystrom): This is a pretty big hack. Up until now, every language
-  // version was a strict syntactic superset of all previous versions. That let
-  // the formatter parse every file at the latest language version without
-  // having to detect each file's actual version, which requires digging around
-  // in the file system for package configs and looking for "@dart" comments in
-  // files. It also means the library API that parses arbitrary strings doesn't
-  // have to worry about what version the code should be interpreted as.
-  //
-  // But with patterns, a small number of switch cases are no longer
-  // syntactically valid. Breakage from this is very rare. Instead of adding
-  // the machinery to detect language versions (which is likely to be slow and
-  // brittle), we just try parsing everything with patterns enabled. When a
-  // parse error occurs, we try parsing it again with pattern disabled. If that
-  // happens to parse without error, then we use that result instead.
-  ParseStringResult _parse(String source, String? uri,
-      {required bool patterns}) {
-    var version = patterns ? Version(3, 3, 0) : Version(2, 19, 0);
-
+  /// Parse [source] from [uri] at language [version].
+  ParseStringResult _parse(String source, String? uri, Version version) {
     // Don't pass the formatter's own experiment flag to the parser.
     var experiments = experimentFlags.toList();
     experiments.remove(tallStyleExperimentFlag);